+++ libs/dot-components/src/lib/components/breadcrumbs/Breadcrumbs.spec.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import userEvent from '@testing-library/user-event';
-import { render, screen } from '../../testing-utils';
+import { render, screen, waitFor } from '../../testing-utils';
 import { BreadcrumbProps, BreadcrumbItem, DotBreadcrumbs } from './Breadcrumbs';
 
 describe('Breadcrumbs', () => {
@@ -14,14 +14,17 @@ describe('Breadcrumbs', () => {
     { href: '#', text: 'Link 1' },
     { href: '#', text: 'Link 2' },
     { href: '#', text: 'Link 3' },
+    { href: '#', text: 'Link 4' },
   ];
 
   it('should have unchanged API', () => {
     const props = {
+      expansionMenu: false,
       items: dummyItems,
       maxItems: 5,
     };
     const breadcrumbProps: BreadcrumbProps = {
+      expansionMenu: false,
       items: dummyItems,
       maxItems: 5,
     };
@@ -38,6 +41,24 @@ describe('Breadcrumbs', () => {
     const links = screen.getAllByRole('link');
 
     expect(links.length).toEqual(2);
+  });
+
+  it('should show expansion menu on ... click when more than maxItems and expansionMenu is true', () => {
+    render(
+      <DotBreadcrumbs
+        items={dummyItemsNoOnClick}
+        expansionMenu={true}
+        maxItems={2}
+      />
+    );
+    waitFor(() => {
+      expect(screen.getByText('Link 2')).not.toBeVisible();
+    });
+    const expandButton = screen.getByRole('button');
+    userEvent.click(expandButton);
+    waitFor(() => {
+      expect(screen.getByText('Link 2')).toBeVisible();
+    });
   });
 
   it('should call onClick if one is passed down as a prop', () => {